corgi: Rel emits a byte mask, now that the host widens it
Reinstates what `d744eef` reverted. The core's mask idiom is "nonzero is true" and leaves
are width-tagged, so one bit was being stored in eight bytes; every consumer already reads
any width. What was missing was not in corgi — it was DDIR treating a comparison's result
as a u64 VALUE, which `Shape::Prim` carrying the leaf width then made a type difference,
which DDIR's own `Eq` lowering folded to a constant. Three programs silently returned
nothing.

The rule that fixes it belongs to the host, and DDIR now has it: widen a mask where it
becomes a value, not where it is made. It has to be the host's, because an order-preserving
widen has to know the KIND — `Rel` is kind-blind on stored bytes that are order-preserving
at their own width, so corgi cannot widen one correctly, and DDIR can, since it emits
`ToSigned` itself. See `dd-corgi-dist`'s `corgi::logic::width`.

Two ways to read a mask, and the split is by ACCESS PATTERN rather than by taste.
`filter_mask` is generic over the width and materializes nothing — used where the loop is
the mask's alone (`Filter`, `arrange::mask_positions`), and it is what this branch keeps
from the revert instead of going back to `as_byte_mask` there. `as_byte_mask` returns a
`Cow` and is for a consumer that cannot dispatch above its own loop: `blend` hands the mask
down a structural recursion to one loop per leaf field, and `try_filter` reads it beside an
error column it is building. With a byte mask the `Cow` borrows, so that path now costs
nothing where before it narrowed a u64 column. `tags_usize` is the discriminant sibling,
widening rather than narrowing because a tag is an index and not a truth value.

Measured (perf-gaps.md, 1 M rows): B1 filter_values 1.65 -> 1.34 ns/row, B2 cmp_select
2.13 -> 1.60. Verified: 143 tests, clippy -D warnings, the tour, and DDIR's 33 AoC parts at
32/33 on both backends — the same one failure as before this branch (day13 p1's `if` with
mismatched branch shapes), where the byte mask alone broke 30 of 33.

+++ corgi/src/engine.rs
@@ -43,7 +43,8 @@ mod generators {
     ///
     /// The mask's WIDTH is the producer's choice — the core's idiom is "nonzero is true" and a leaf
     /// is width-tagged — so this dispatches it once, above the pass, rather than making the caller
-    /// normalize (which would be a pass over the mask to save nothing).
+    /// normalize (which would be a pass over the mask to save nothing). `Rel` produces a byte mask,
+    /// so the `U8` arm is the common one and no other arm costs anything to have.
     pub(crate) fn filter_mask(bounds: &Bounds, mask: &Prim) -> (Vec<usize>, Vec<usize>) {
         fn scan<T: Copy + Default + PartialEq>(bounds: &Bounds, m: &[T]) -> (Vec<usize>, Vec<usize>) {
             let mut idx = Vec::new();
@@ -272,7 +273,7 @@ pub(crate) fn gather_lanes(srcs: &[Option<&Value>], tags: &[usize], off: &[usize
 /// position) has no constant slot to blend into, so it falls back to the two-source [`gather_lanes`]
 /// — the same split `scatter` makes, and for the same reason. The split is per LEVEL, not per value:
 /// a product blends each leaf field directly and only gathers the fields that need it.
-pub(crate) fn blend(mask: &[u64], then: Value, els: Value) -> Value {
+pub(crate) fn blend(mask: &[u8], then: Value, els: Value) -> Value {
     match (then, els) {
         (Value::Prim(t), Value::Prim(e)) => Value::Prim(t.blend(e, mask)),
         (Value::Prod(ts), Value::Prod(es)) => {
@@ -390,12 +391,12 @@ mod tests {
     #[test]
     fn blend_matches_the_gather_lanes_path() {
         // the general path, written out: row i from lane `mask[i] != 0`, at its own position.
-        fn oracle(mask: &[u64], then: &Value, els: &Value) -> Value {
+        fn oracle(mask: &[u8], then: &Value, els: &Value) -> Value {
             let tags: Vec<usize> = mask.iter().map(|&m| (m != 0) as usize).collect();
             let off: Vec<usize> = (0..tags.len()).collect();
             gather_lanes(&[Some(els), Some(then)], &tags, &off)
         }
-        let mask = [1u64, 0, 0, 1];
+        let mask = [1u8, 0, 0, 1];
         let list = |ends: Vec<usize>, xs: &[u64]| Value::List(ends.into(), Box::new(u(xs)));
 
         // leaf, product of leaves, unit — the lane-wise path.
